stmtTree: Insert non-conditional node as early as possible
Prev impl just insert non-conditional to the tail of a STMT_SEQ when
generating stmtTree from expTree.
This commit try to insert the node to the earliest possible position in
a STMT_SEQ to minimize dependency path. This is accomplished by
elevating existing STMT_NODE to STMT_SEQ then append to it. Allowing
much more extensive conditional operations merging